% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_pts_frst.R
\name{assign_pts_frst}
\alias{assign_pts_frst}
\title{Assign randomly sampled points to their nearest grid based on the difference
between the polygons original and desired size.}
\usage{
assign_pts_frst(x, props, nf_pct)
}
\arguments{
\item{x}{distance matrix between the points and their potential neighbors polygons}

\item{props}{the existing proportional area of each potential neighbor polygons}

\item{nf_pct}{the desired proportional area of each potential neighbor polygon}
}
\description{
Assign randomly sampled points to their nearest grid based on the difference
between the polygons original and desired size.
}
\keyword{internal}
