% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignGrid_pts.R
\name{assignGrid_pts}
\alias{assignGrid_pts}
\title{place random points in the polygon which will be dissolved with the larger polygons}
\usage{
assignGrid_pts(neighb_grid, focal_grid, props, nf_pct)
}
\arguments{
\item{neighb_grid}{}

\item{focal_grid}{}

\item{props}{}

\item{nf_pct}{}
}
\description{
This function is ran within \code{GridBasedSample} it will place points throughout polygon geometries
which should be merged to larger polygons and assign them to neighboring polygons
based on how much area we want to grow these polygons too.
}
\keyword{internal}
