% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_neigh_directions.R
\name{first_neigh_directions}
\alias{first_neigh_directions}
\title{We use spdep::knearneigh to ensure we obtain 4 of the nearest neighbors to a polygon}
\usage{
first_neigh_directions(from, destinations)
}
\arguments{
\item{from}{a point on surface of the grid which will be merged. POS will ensure it lands on a feature.}

\item{destinations}{the set of kearneigh which are accepting polygon merges}
}
\description{
However, this system has limitations, if an island is ringed by land, it would work adequately.
But when the island is off a coast, it may return neighbors which are 'behind' another neighbor.
We draw lines from the focal portion of the grid we will be combining to each neighbor,
if the lines cross any other neighbor they are discarded.
}
\details{
This is preferable to simply using the nearest feature (e.g. sf:;st_nearest_feature) when the chain of islands is elongated and may more appropriately be split across multiple grids downstream from here.
}
\keyword{internal}
