% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointBasedSample.R
\name{get_elements}
\alias{get_elements}
\title{Recursively grab a named component of a list.}
\usage{
get_elements(x, element)
}
\arguments{
\item{x}{a list of lists}

\item{element}{the quoted name of the list element to extract.}
}
\description{
Recursively grab a named component of a list.
}
