% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeDispersion.R
\name{greatCircleDistance}
\alias{greatCircleDistance}
\title{Haversine Distance Calculation}
\usage{
greatCircleDistance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Double. column holding coords of 'focal' population}

\item{lon1}{Double. column holding coords of 'focal' population}

\item{lat2}{Double. column holding coords of 'non-focal' population}

\item{lon2}{Double. column holding coords of 'non-focal' population}
}
\description{
Calculate geographic distances on the geoid.
This results in more accurate distance calculations than a planar system.
Function is mostly used internally by \code{maximize_dispersion}
}
\details{
calculate distances between sites (Haversine formula)
}
\examples{
n_sites <- 5 # number of known populations
 df <- data.frame(
   site_id = seq_len(n_sites),
   lat = runif(n_sites, 25, 30), 
   lon = runif(n_sites, -125, -120)
 )

dist_mat <- sapply(1:nrow(df), function(i) {
   greatCircleDistance(
     df$lat[i], df$lon[i],
     df$lat, df$lon
   )
 })
#head(dist_mat)
}
