% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDBasedSample.R
\name{IBDBasedSample}
\alias{IBDBasedSample}
\title{Sample a species based on Isolation by Geographic Distance (IBD)}
\usage{
IBDBasedSample(
  x,
  n,
  fixedClusters,
  n_pts,
  template,
  prop_split,
  min.nc,
  max.nc
)
}
\arguments{
\item{x}{a Raster surface to sample the points within, e.g. the output of SDM$Supplemented.}

\item{n}{Numeric. the number of clusters desired.}

\item{fixedClusters}{Boolean. Defaults to TRUE, which will create n clusters. If False then use NbClust::NbClust to determine the optimal number of clusters.}

\item{n_pts}{Numeric. the number of points to use for generating the clusters, these will be randomly sampled within the mask area \code{mask}. Defaults to 1000, which generally allows for enough points to be split for the KNN training.}

\item{template}{Raster. A raster file which can be used as a template for plotting.}

\item{prop_split}{Numeric. The proportion of records to be used for training the KNN classifier. Defaults to
0.8 to use 80\% of records for training and 20\% for the independent test sample.}

\item{min.nc}{Numeric. Minimum number of clusters to test if fixedClusters=FALSE, defaults to 5.}

\item{max.nc}{Numeric. Maximum number of clusters to test if fixedClusters=FALSE, defaults to 20.}
}
\value{
An simple features (sf) object containing the final grids for saving to computer. See the vignette for questions about saving the two main types of spatial data models (vector - used here, and raster).
}
\description{
Create n seed collection areas based on the distance geographic (great circle) distance between points.
}
