% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBRBasedSample.R
\name{IBRBasedSample}
\alias{IBRBasedSample}
\title{Sample a species based on Isolation by Resistance Distance (IBR)}
\usage{
IBRBasedSample(
  base_raster,
  pop_raster,
  resistance_surface,
  pts_sf,
  ibr_matrix,
  fixedClusters = FALSE,
  n = NULL,
  min.nc = 5,
  max.nc = 20,
  planar_proj,
  max_geo_cells = NULL,
  geo_iters = 10,
  boundary_width = 2,
  distance_method = c("haversine", "cosine")
)
}
\arguments{
\item{base_raster}{a Raster surface to sample the points within, ....}

\item{pop_raster}{Spatiall buffered population raster from \code{populationResistance}.}

\item{resistance_surface}{a resistance surface raster from  \code{build_resistance_surface} or \code{populationResistance}.}

\item{pts_sf}{sf/tibble/dataframe of point locations from \code{populationResistance} $pts_sf}

\item{ibr_matrix}{Distance matrix, from \code{populationResistance} slot $Rmat}

\item{fixedClusters}{Boolean. Defaults to FALSE, which will create n clusters. If False then use NbClust::NbClust to determine the optimal number of clusters. TRUE not yet suported.}

\item{n}{Numeric. the number of clusters desired.}

\item{min.nc}{Numeric. Minimum number of clusters to test if fixedClusters=FALSE, defaults to 5.}

\item{max.nc}{Numeric. Maximum number of clusters to test if fixedClusters=FALSE, defaults to 20.}

\item{planar_proj}{For numbering clusters in return object.}

\item{geo_iters}{Numeric. Maximum number of iterations to grow hulls by.}

\item{boundary_width}{Numeric. Maximum boundary width before switching to final assignment method.}
}
\description{
Create \code{n} seed collection areas based on the habitat, and geographic, resistance between points.
}
\examples{
 # see package vignette
}
