% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrioritizeSample.R
\name{order_by_distance_variance}
\alias{order_by_distance_variance}
\title{Order zones by minimizing distance variance}
\usage{
order_by_distance_variance(x, metric = c("var", "sd", "energy", "cv"))
}
\arguments{
\item{x}{an sf object containing the zones to be ordered}

\item{metric}{character. The metric to minimize when ordering zones. Options are "var" (variance), "sd" (standard deviation), "energy" (sum of squared distances), and "cv" (coefficient of variation).}
}
\value{
A numeric vector representing the order of zones based on the specified distance variance metric
}
\description{
Order a set of spatial zones based on minimizing distance variance
}
\keyword{internal}
