% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointBasedSample.R
\name{PointBasedSample}
\alias{PointBasedSample}
\title{Generate a sampling grid based off of regularly sampled points across the species range.}
\usage{
PointBasedSample(polygon, n, collections, reps, BS.reps)
}
\arguments{
\item{polygon}{the input sf polygon, i.e. species range or administrative unit, where sampling is desired.}

\item{n}{Numeric. The total number of desired collections. Defaults to 20.}

\item{collections}{an sf point geometry data set of where existing collections have been made.}

\item{reps}{further arguments passed to np.boot}

\item{BS.reps}{number of bootstrap replicates for evaluating results.}
}
\value{
A list containing two objects, the first the results of bootstrap simulations.
The second an sf dataframe containing the polygons with the smallest amount of variance in size.
}
\description{
This function utilizes a regular, or nearly so in the case of existing collections, grid of points
to develop a sampling scheme or n polygons.
}
\examples{
#' Utilize a grid based stratified sample for drawing up polygons
ri <- spData::us_states |>
  dplyr::select(NAME) |>
  dplyr::filter(NAME == 'Rhode Island') |>
  sf::st_transform(32617)
  
 system.time(
  out <- PointBasedSample(polygon = ri, reps = 10, BS.reps = 10) # set very low for example
 )
# the function is actually very fast; 150 voronoi reps, with 9999 BS should only take about
# 2 seconds per species so not much concern on the speed end of things!
head(out$SummaryData)
plot(out$Geometry)

}
