% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceFinalGrids.R
\name{reduceFinalGrids}
\alias{reduceFinalGrids}
\title{More sliver fixing}
\usage{
reduceFinalGrids(final_grids)
}
\arguments{
\item{final_grids}{truthfully nearly final at this point.}
}
\description{
in some instances tiny little grids are tagged along from processing.
we will just give these to an arbitrary nearest feature. They would be
annoying if by chance (... they are about a millionth of the area...)
an arbitrary random point was in them and missed,
but the areas themselves tend to be remarkable inconsequential, and worth randomly
reassigning to a neighbor
}
\keyword{internal}
