% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescaleRasters.R
\name{RescaleRasters}
\alias{RescaleRasters}
\title{Rescale a raster stack to reflect the beta coefficients from a glmnet model}
\usage{
RescaleRasters(model, predictors, training_data, sdModel, pred_mat)
}
\arguments{
\item{model}{the final output model from glmnet from \code{elasticSDM}}

\item{predictors}{the raster stack to use for the process from \code{elasticSDM}}

\item{training_data}{the same data that went into the glmnet model, this is used
for calculating variance which is required for the scaling process. From \code{elasticSDM}}

\item{the}{Prediction matrix from \code{elasticSDM}}
}
\description{
These rescaled rasters can then be used for clustering, and predicting the
results of cluster analysis back into space for a final product.
}
