% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeHavens-package.R
\docType{package}
\name{safeHavens-package}
\alias{safeHavens}
\alias{safeHavens-package}
\title{safeHavens: Developing sampling schemas for ex situ conservation of plant germplasm}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package helps germplasm curators communicate areas of interest to collection teams for them to collect new material for accession. It provides six different sampling approaches for curators to choose from for each individual taxon they hope to process. It also allows for easy integration into existing workflows and will put out the required spatial data to share with collection teams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sagesteppe.github.io/safeHavens/}
}

}
\author{
\strong{Maintainer}: Reed Benkendorf \email{rbenkendorf@chicagobotanic.org} (\href{https://orcid.org/0000-0001-9178-3087}{ORCID})

}
