% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapGrids.R
\name{snapGrids}
\alias{snapGrids}
\title{turn the point grid suggestions made by assignGrid_pts into polygons}
\usage{
snapGrids(x, neighb_grid, focal_grid)
}
\arguments{
\item{x}{output of \code{assignGrid_pts}}

\item{neighb_grid}{the neighboring grid options.}

\item{focal_grid}{the grid to reassign the area of.}
}
\description{
This function is part of the grid based sampling process to turn small grid cells,
which are to be broken up, into a larger existing grid cells.
}
