% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healPolygons.R
\name{snapR}
\alias{snapR}
\title{Clean up unioned geometries - part 2}
\usage{
snapR(x)
}
\arguments{
\item{x}{the output of healPolygons}
}
\description{
this function uses sf::st_snap to remove small lines and other artifacts associated
with the unioning of polygons
}
\keyword{internal}
