% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonBasedSample.R
\name{split_cols}
\alias{split_cols}
\title{split and extract the temperature values from Tmin and AHM columns}
\usage{
split_cols(dat, y, sep = "-")
}
\arguments{
\item{dat}{data frame with the columns required to split.}

\item{y}{character. column name to split.}

\item{sep}{character. separator between the values to split on. Default is '-'.}
}
\description{
Programmed for the Bower provisional seed zone products, a helper function for
separating and recovering the values from the columns in the data.
}
\examples{
df = data.frame(
  'Tmin_class' = c('10 - 15 Deg. F.', '15 - 20 Deg. F.', '> 55 Deg. F.' ),
  'AHM_class' = c('2 - 3', '6 - 12', '3 - 6')
)
split_cols(df, 'Tmin_class')
split_cols(df, 'AHM_class')

}
