% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestGridSizes.R
\name{TestGridSizes}
\alias{TestGridSizes}
\title{Get an estimate for how many grids to draw over a species range}
\usage{
TestGridSizes(target)
}
\arguments{
\item{target}{a species range as a simple feature (\code{sf}) object.}
}
\value{
A dataframe with testing results for each grid combination.
A user needs to select the optimal grid size based on a tradeoff with minimizing variance, without creating too many grids which will need to be erased.
In the Rhode Island example I would use the 'Original' option which asks for 4 x grids and 7 y grids.
}
\description{
This function uses the dimensions of a species grid to estimate how many grids
would need to be added in the x and y directions to cover it with 20 grid cells
of roughly equal areas.
}
\examples{
ri <- spData::us_states |>
dplyr::select(NAME) |>
   dplyr::filter(NAME == 'Rhode Island') |>
   sf::st_transform(32617)

sizeOptions <- TestGridSizes(ri)
head(sizeOptions)
}
