% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainKNN.R
\name{trainKNN}
\alias{trainKNN}
\title{partition data and train a simple KNN model}
\usage{
trainKNN(x, split_prop)
}
\arguments{
\item{x}{the weighted matrix including a class ID in the column 'ID'}

\item{split}{prop for the data partitions.}
}
\description{
Simply use this to partition your test data and quickly train a simple model
}
