% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeHavens.r
\docType{data}
\name{WesternEcoregions}
\alias{WesternEcoregions}
\title{Biogeographic regions of the Neotropics.}
\format{
A spatial vector data set.
}
\description{
A vector data set of Bioregions developed by Morrone et al. 2022
Downloaded from https://neotropicalmap.atlasbiogeografico.com/ and
simplified using mapshaper.
}
\details{
\itemize{
\item Provincias Full name of the bioregion
}
}
\keyword{datasets}
